Move `DepthwiseSeparableConvModule` into `conv_module`
