+++ src/otx/algo/detection/backbones/csp_darknet.py
@@ -18,8 +18,7 @@ from torch.nn.modules.batchnorm import _BatchNorm
 from otx.algo.common.layers import SPPBottleneck
 from otx.algo.detection.layers import CSPLayer
 from otx.algo.modules.base_module import BaseModule
-from otx.algo.modules.conv_module import Conv2dModule
-from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
+from otx.algo.modules.conv_module import Conv2dModule, DepthwiseSeparableConvModule
 
 
 class Focus(nn.Module):
