+++ src/otx/algo/detection/necks/cspnext_pafpn.py
@@ -17,8 +17,7 @@ from torch import Tensor, nn
 
 from otx.algo.detection.layers import CSPLayer
 from otx.algo.modules.base_module import BaseModule
-from otx.algo.modules.conv_module import Conv2dModule
-from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
+from otx.algo.modules.conv_module import Conv2dModule, DepthwiseSeparableConvModule
 
 
 class CSPNeXtPAFPN(BaseModule):
