+++ src/otx/algo/modules/__init__.py
@@ -5,8 +5,7 @@
 """This module implementation is a code implementation copied or replaced from mmcv.cnn.bricks."""
 
 from .activation import build_activation_layer
-from .conv_module import Conv2dModule, Conv3dModule
-from .depthwise_separable_conv_module import DepthwiseSeparableConvModule
+from .conv_module import Conv2dModule, Conv3dModule, DepthwiseSeparableConvModule
 from .norm import FrozenBatchNorm2d, build_norm_layer
 from .padding import build_padding_layer
