+++ src/otx/algo/modules/conv_module.py
@@ -10,11 +10,12 @@ from __future__ import annotations
 
 import warnings
 from functools import partial
-from typing import TYPE_CHECKING, TypeVar
+from typing import TypeVar
 
 import torch
-from torch import nn
+from torch import Tensor, nn
 from torch.nn.modules.batchnorm import _BatchNorm as BatchNorm
+from torch.nn.modules.conv import _ConvNd as ConvNd
 from torch.nn.modules.instancenorm import _InstanceNorm as InstanceNorm
 
 from otx.algo.utils.weight_init import constant_init, kaiming_init
@@ -23,14 +24,10 @@ from .activation import build_activation_layer
 from .norm import build_norm_layer
 from .padding import build_padding_layer
 
-if TYPE_CHECKING:
-    from torch.nn.modules.conv import _ConvNd as ConvNd
+T_ConvND = TypeVar("T_ConvND", bound=ConvNd)
 
 
-T_ConvND = TypeVar("T_ConvND", bound=nn.Conv1d | nn.Conv2d | nn.Conv3d | nn.ConvTranspose2d | nn.ConvTranspose3d)
-
-
-def efficient_conv_bn_eval_forward(bn: BatchNorm, conv: ConvNd, x: torch.Tensor) -> torch.Tensor:
+def efficient_conv_bn_eval_forward(bn: BatchNorm, conv: T_ConvND, x: Tensor) -> Tensor:
     """Implementation based on https://arxiv.org/abs/2305.11624.
 
     "Tune-Mode ConvBN Blocks For Efficient Transfer Learning"
@@ -41,8 +38,8 @@ def efficient_conv_bn_eval_forward(bn: BatchNorm, conv: ConvNd, x: torch.Tensor)
 
     Args:
         bn (_BatchNorm): a BatchNorm module.
-        conv (nn._ConvNd): a conv module
-        x (torch.Tensor): Input feature map.
+        conv (T_ConvND): a conv module
+        x (Tensor): Input feature map.
     """
     # These lines of code are designed to deal with various cases
     # like bn without affine transform, and conv without bias
@@ -256,16 +253,16 @@ class ConvModule(nn.Module):
         if self.with_norm:
             constant_init(self.norm_layer, 1, bias=0)
 
-    def forward(self, x: torch.Tensor, activate: bool = True, norm: bool = True) -> torch.Tensor:
+    def forward(self, x: Tensor, activate: bool = True, norm: bool = True) -> Tensor:
         """Forward pass of the ConvModule.
 
         Args:
-            x (torch.Tensor): Input tensor.
+            x (Tensor): Input tensor.
             activate (bool, optional): Whether to apply activation. Defaults to True.
             norm (bool, optional): Whether to apply normalization. Defaults to True.
 
         Returns:
-            torch.Tensor: Output tensor.
+            Tensor: Output tensor.
         """
         layer_index = 0
         while layer_index < len(self.order):
@@ -358,6 +355,105 @@ class ConvModule(nn.Module):
         self.turn_on_efficient_conv_bn_eval(efficient_conv_bn_eval)
 
         return self
+
+
+class DepthwiseSeparableConvModule(nn.Module):
+    """Depthwise separable convolution module.
+
+    See https://arxiv.org/pdf/1704.04861.pdf for details.
+
+    This module can replace a ConvModule with the conv block replaced by two
+    conv block: depthwise conv block and pointwise conv block. The depthwise
+    conv block contains depthwise-conv/norm/activation layers. The pointwise
+    conv block contains pointwise-conv/norm/activation layers. It should be
+    noted that there will be norm/activation layer in the depthwise conv block
+    if `norm_cfg` and `act_cfg` are specified.
+
+    Args:
+        in_channels (int): Number of channels in the input feature map.
+            Same as that in ``nn._ConvNd``.
+        out_channels (int): Number of channels produced by the convolution.
+            Same as that in ``nn._ConvNd``.
+        kernel_size (int | tuple[int]): Size of the convolving kernel.
+            Same as that in ``nn._ConvNd``.
+        stride (int | tuple[int]): Stride of the convolution.
+            Same as that in ``nn._ConvNd``. Default: 1.
+        padding (int | tuple[int]): Zero-padding added to both sides of
+            the input. Same as that in ``nn._ConvNd``. Default: 0.
+        dilation (int | tuple[int]): Spacing between kernel elements.
+            Same as that in ``nn._ConvNd``. Default: 1.
+        norm_cfg (dict): Default norm config for both depthwise ConvModule and
+            pointwise ConvModule. Default: None.
+        act_cfg (dict): Default activation config for both depthwise ConvModule
+            and pointwise ConvModule. Default: dict(type='ReLU').
+        dw_norm_cfg (dict): Norm config of depthwise ConvModule. If it is
+            None, it will be the same as `norm_cfg`. Default: None.
+        dw_act_cfg (dict): Activation config of depthwise ConvModule. If it is
+            None, it will be the same as `act_cfg`. Default: None.
+        pw_norm_cfg (dict): Norm config of pointwise ConvModule. If it is
+            None, it will be the same as `norm_cfg`. Default: None.
+        pw_act_cfg (dict): Activation config of pointwise ConvModule. If it is
+            None, it will be the same as `act_cfg`. Default: None.
+        kwargs (optional): Other shared arguments for depthwise and pointwise
+            ConvModule. See ConvModule for ref.
+    """
+
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        kernel_size: int | tuple[int, int],
+        stride: int | tuple[int, int] = 1,
+        padding: int | tuple[int, int] = 0,
+        dilation: int | tuple[int, int] = 1,
+        norm_cfg: dict | None = None,
+        act_cfg: dict | None = None,
+        dw_norm_cfg: dict | None = None,
+        dw_act_cfg: dict | None = None,
+        pw_norm_cfg: dict | None = None,
+        pw_act_cfg: dict | None = None,
+        **kwargs,
+    ):
+        if act_cfg is None:
+            act_cfg = {"type": "ReLU"}
+
+        super().__init__()
+        assert "groups" not in kwargs, "groups should not be specified"  # noqa: S101
+
+        # if norm/activation config of depthwise/pointwise Conv2dModule is not
+        # specified, use default config.
+        dw_norm_cfg = dw_norm_cfg or norm_cfg
+        dw_act_cfg = dw_act_cfg or act_cfg
+        pw_norm_cfg = pw_norm_cfg or norm_cfg
+        pw_act_cfg = pw_act_cfg or act_cfg
+
+        # depthwise convolution
+        self.depthwise_conv = Conv2dModule(
+            in_channels,
+            in_channels,
+            kernel_size,
+            stride=stride,
+            padding=padding,
+            dilation=dilation,
+            groups=in_channels,
+            norm_cfg=dw_norm_cfg,
+            act_cfg=dw_act_cfg,
+            **kwargs,
+        )
+
+        self.pointwise_conv = Conv2dModule(
+            in_channels,
+            out_channels,
+            1,
+            norm_cfg=pw_norm_cfg,
+            act_cfg=pw_act_cfg,
+            **kwargs,
+        )
+
+    def forward(self, x: Tensor) -> Tensor:
+        """Forward."""
+        x = self.depthwise_conv(x)
+        return self.pointwise_conv(x)
 
 
 class Conv2dModule(ConvModule):
