+++ tests/unit/algo/detection/necks/test_yolox_pafpn.py
@@ -8,7 +8,7 @@ Reference : https://github.com/open-mmlab/mmdetection/blob/v3.2.0/tests/test_mod
 
 import torch
 from otx.algo.detection.necks.yolox_pafpn import YOLOXPAFPN
-from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
+from otx.algo.modules.conv_module import DepthwiseSeparableConvModule
 
 
 class TestYOLOXPAFPN:
